bind lifcycle fix

+++ app/src/main/java/appliedlife/pvtltd/SHEROES/views/fragments/FeedFragment.java
@@ -1196,7 +1196,7 @@ public class FeedFragment extends BaseFragment implements IFeedView, FeedItemCal
         if (getActivity() == null || getActivity().isFinishing()) return;
         PopupMenu popup = new PopupMenu(getActivity(), view);
         if (null != mUserPreference && mUserPreference.isSet() && null != mUserPreference.get().getUserSummary()) {
-            if (articleObj.getCreatedBy() == (long)mUserPreference.get().getUserSummary().getUserId()) {
+            if (articleObj.getCreatedBy() ==mUserPreference.get().getUserSummary().getUserId()) {
                 popup.getMenu().add(0, R.id.edit, 1, menuIconWithText(getResources().getDrawable(R.drawable.ic_create), getResources().getString(R.string.ID_EDIT)));
                 popup.getMenu().add(0, R.id.delete, 2, menuIconWithText(getResources().getDrawable(R.drawable.ic_delete), getResources().getString(R.string.ID_DELETE)));
             }
